factored out lib functions, added cli script
